Foco Heatmapping
This is my progress so far to use the latitude longitude data from scooters in Fort Collins to develop a heatmap of popular start/end locations and predict future spots. My progress has led me to learning how to use real life mapping data to create the map of Fort Collins, and apply the general application of the latitude, longitude data into a geographic plot of start points and a heatmap of overall density in Fort Collins. Next process will be implementing a model like nearest neighbors to implement a more specialized version of our mapping model.